User article

+++ article/views.py
@@ -1,39 +1,54 @@
-from django.shortcuts import render,redirect
+from django.shortcuts import render, redirect
 
 from .models import ArticlePost
 import markdown
 from django.http import HttpResponse
 from .forms import ArticlePostForm
 from django.contrib.auth.models import User
-
-
-
+from django.contrib.auth.decorators import login_required
+from django.core.paginator import Paginator
 
 
 def article_list(request):
-    articles = ArticlePost.objects.all()
+    if request.GET.get('order') == 'total_views':
+        article_list = ArticlePost.objects.all().order_by('-total_views')
+        order = 'total_views'
+    else:
+        article_list = ArticlePost.objects.all()
+        order = 'normal'
+
+    # 每页显示 3 篇文章
+    paginator = Paginator(article_list, 3)
+    page = request.GET.get('page')
+    articles = paginator.get_page(page)
+
     # 需要传递给模板的上下文
-    context = {'articles': articles}
+    context = {'articles': articles, 'order': order}
     return render(request, 'article/list.html', context)
 
 
 def article_detail(request, id):
     article = ArticlePost.objects.get(id=id)
+
+    # 浏览量 +1
+    article.total_views += 1
+    article.save(update_fields=['total_views'])
+
     article.body = markdown.markdown(article.body,
-    extensions=['markdown.extensions.extra',
-    'markdown.extensions.codehilite'])
+                    extensions=['markdown.extensions.extra',
+                                'markdown.extensions.codehilite'])
 
     context = {'article': article}
     return render(request, 'article/detail.html', context)
 
 
-
+@login_required(login_url='/userprofile/login/')
 def article_create(request):
     if request.method == 'POST':
-        article_post_form=ArticlePostForm(data=request.POST)
+        article_post_form = ArticlePostForm(data=request.POST)
         if article_post_form.is_valid():
             new_article = article_post_form.save(commit=False)
-            new_article.author=User.objects.get(id=1)
+            new_article.author = User.objects.get(id=request.user.id)
             new_article.save()
             return redirect('article:list')
         else:
@@ -41,17 +56,49 @@ def article_create(request):
     else:
         article_post_form = ArticlePostForm()
         context = {'article_post_form': article_post_form}
-        return render(request, 'article/create.html',context)
+        return render(request, 'article/create.html', context)
 
+
+@login_required(login_url='/userprofile/login/')
 def article_delete(request, id):
-    article=ArticlePost.objects.get(id=id)
+    article = ArticlePost.objects.get(id=id)
     article.delete()
     return redirect('article:list')
 
+
+@login_required(login_url='/userprofile/login/')
 def article_safe_delete(request, id):
+    article = ArticlePost.objects.get(id=id)
+
+    # 过滤非作者的用户
+    if request.user != article.author:
+        return HttpResponse("抱歉，你无权删除这篇文章。")
+
     if request.method == 'POST':
-        article = ArticlePost.objects.get(id=id)
         article.delete()
         return redirect('article:list')
     else:
         return HttpResponse('请求错误')
+
+
+@login_required(login_url='/userprofile/login/')
+def article_update(request, id):
+    article = ArticlePost.objects.get(id=id)
+
+    # 过滤非作者的用户
+    if request.user != article.author:
+        return HttpResponse("抱歉，你无权修改这篇文章。")
+
+    if request.method == 'POST':
+        article_post_form = ArticlePostForm(data=request.POST)
+        if article_post_form.is_valid():
+            article.title = request.POST['title']
+            article.body = request.POST['body']
+            article.save()
+            return redirect('article:detail', id=id)
+        else:
+            return HttpResponse('表单内容有误，请重新填写')
+    else:
+        article_post_form = ArticlePostForm()
+        context = {'article': article, 'article_post_form': article_post_form}
+        return render(request, 'article/update.html', context)
